+++ src/main/kotlin/com/atlassian/performance/tools/dockerinfrastructure/api/jira/Jira.kt
@@ -4,5 +4,4 @@ import java.net.URI
 
 interface Jira : AutoCloseable {
     fun getUri(): URI
-    fun getDockerUri(): URI
 }