+++ CHANGELOG.md
@@ -23,8 +23,14 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/docker-infrastructure/branches/compare/master%0Drelease-0.1.1
 
-## Fixed
+### Fixed
 - Provision jira with a default dataset.
+
+### Added
+- `JiraCoreFormula` can build `Jira` that is accessible from docker or host.
+
+### Removed
+- Remove `getDockerUri` method from `Jira` interface.
 
 ## [0.1.1] - 2019-01-10
 [0.1.1]: https://bitbucket.org/atlassian/docker-infrastructure/branches/compare/release-0.1.1%0Drelease-0.1.0